+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Add new project management commands: `projects:create`, `projects:list`.
 * Add new app management commands: `apps:create`, `apps:list`, `apps:sdkconfig`.
 * Improve `init` command to be able to create a new project.
+* Automatically choose a port for the Firestore emulator to serve WebChannel traffic.